+++ utils/wrappers.py
@@ -9,13 +9,24 @@ class DoneOnSuccessWrapper(gym.Wrapper):
     Useful for GoalEnv.
     """
 
-    def __init__(self, env, reward_offset=1.0):
+    def __init__(self, env, reward_offset=1.0, n_successes=1):
         super(DoneOnSuccessWrapper, self).__init__(env)
         self.reward_offset = reward_offset
+        self.n_successes = n_successes
+        self.current_successes = 0
+
+    def reset(self):
+        self.current_successes = 0
+        return self.env.reset()
 
     def step(self, action):
         obs, reward, done, info = self.env.step(action)
-        done = done or info.get("is_success", False)
+        if info.get("is_success", False):
+            self.current_successes += 1
+        else:
+            self.current_successes = 0
+        # number of successes in a row
+        done = done or self.current_successes >= self.n_successes
         reward += self.reward_offset
         return obs, reward, done, info
 
@@ -81,6 +92,64 @@ class TimeFeatureWrapper(gym.Wrapper):
         return np.concatenate((obs, [time_feature]))
 
 
+class TimeFeatureObsDictWrapper(gym.Wrapper):
+    """
+    Add remaining time to observation space for fixed length episodes.
+    See https://arxiv.org/abs/1712.00378 and https://github.com/aravindr93/mjrl/issues/13.
+
+    :param env: (gym.Env)
+    :param max_steps: (int) Max number of steps of an episode
+        if it is not wrapped in a TimeLimit object.
+    :param test_mode: (bool) In test mode, the time feature is constant,
+        equal to zero. This allow to check that the agent did not overfit this feature,
+        learning a deterministic pre-defined sequence of actions.
+    """
+
+    def __init__(self, env: gym.Env, max_steps: int = 1000, test_mode: bool = False):
+        assert isinstance(env.observation_space, gym.spaces.Dict)
+        # Add a time feature to the observation
+        obs_space = env.observation_space.spaces["observation"]
+        low, high = obs_space.low, obs_space.high
+        low, high = np.concatenate((low, [0])), np.concatenate((high, [1.0]))
+        env.observation_space.spaces["observation"] = gym.spaces.Box(low=low, high=high, dtype=np.float32)
+
+        super(TimeFeatureObsDictWrapper, self).__init__(env)
+
+        try:
+            self._max_steps = env.spec.max_episode_steps
+        except AttributeError:
+            self._max_steps = None
+
+        if self._max_steps is None:
+            self._max_steps = max_steps
+
+        self._current_step = 0
+        self._test_mode = test_mode
+
+    def reset(self):
+        self._current_step = 0
+        return self._get_obs(self.env.reset())
+
+    def step(self, action):
+        self._current_step += 1
+        obs, reward, done, info = self.env.step(action)
+        return self._get_obs(obs), reward, done, info
+
+    def _get_obs(self, obs):
+        """
+        Concatenate the time feature to the current observation.
+
+        :param obs: (np.ndarray)
+        :return: (np.ndarray)
+        """
+        # Remaining time is more general
+        time_feature = 1 - (self._current_step / self._max_steps)
+        if self._test_mode:
+            time_feature = 1.0
+        obs["observation"] = np.concatenate((obs["observation"], [time_feature]))
+        return obs
+
+
 class ActionNoiseWrapper(gym.Wrapper):
     """
     Add gaussian noise to the action (without telling the agent),
@@ -98,6 +167,101 @@ class ActionNoiseWrapper(gym.Wrapper):
         noise = np.random.normal(np.zeros_like(action), np.ones_like(action) * self.noise_std)
         noisy_action = action + noise
         return self.env.step(noisy_action)
+
+
+# from https://docs.obspy.org
+from scipy.signal import iirfilter, sosfilt, zpk2sos
+
+
+def lowpass(data, freq, df, corners=4, zerophase=False):
+    """
+    Butterworth-Lowpass Filter.
+
+    Filter data removing data over certain frequency ``freq`` using ``corners``
+    corners.
+    The filter uses :func:`scipy.signal.iirfilter` (for design)
+    and :func:`scipy.signal.sosfilt` (for applying the filter).
+
+    :type data: numpy.ndarray
+    :param data: Data to filter.
+    :param freq: Filter corner frequency.
+    :param df: Sampling rate in Hz.
+    :param corners: Filter corners / order.
+    :param zerophase: If True, apply filter once forwards and once backwards.
+        This results in twice the number of corners but zero phase shift in
+        the resulting filtered trace.
+    :return: Filtered data.
+    """
+    fe = 0.5 * df
+    f = freq / fe
+    # raise for some bad scenarios
+    if f > 1:
+        f = 1.0
+        msg = "Selected corner frequency is above Nyquist. " + "Setting Nyquist as high corner."
+        print(msg)
+    z, p, k = iirfilter(corners, f, btype="lowpass", ftype="butter", output="zpk")
+    sos = zpk2sos(z, p, k)
+    if zerophase:
+        firstpass = sosfilt(sos, data)
+        return sosfilt(sos, firstpass[::-1])[::-1]
+    else:
+        return sosfilt(sos, data)
+
+
+class LowPassFilterWrapper(gym.Wrapper):
+    """
+    Butterworth-Lowpass
+
+    :param env: (gym.Env)
+    :param freq: Filter corner frequency.
+    :param df: Sampling rate in Hz.
+    """
+
+    def __init__(self, env, freq=5.0, df=25.0):
+        super(LowPassFilterWrapper, self).__init__(env)
+        self.freq = freq
+        self.df = df
+        self.signal = []
+
+    def reset(self):
+        self.signal = []
+        return self.env.reset()
+
+    def step(self, action):
+        self.signal.append(action)
+        filtered = np.zeros_like(action)
+        for i in range(self.action_space.shape[0]):
+            smoothed_action = lowpass(np.array(self.signal)[:, i], freq=self.freq, df=self.df)
+            filtered[i] = smoothed_action[-1]
+        return self.env.step(filtered)
+
+
+class ActionSmoothingWrapper(gym.Wrapper):
+    """
+    Smooth the action using exponential moving average.
+
+    :param env: (gym.Env)
+    :param smoothing_coef: (float) Smoothing coefficient (0 no smoothing, 1 very smooth)
+    """
+
+    def __init__(self, env, smoothing_coef: float = 0.0):
+        super(ActionSmoothingWrapper, self).__init__(env)
+        self.smoothing_coef = smoothing_coef
+        self.smoothed_action = None
+        # from https://github.com/rail-berkeley/softlearning/issues/3
+        # for smoothing latent space
+        # self.alpha = self.smoothing_coef
+        # self.beta = np.sqrt(1 - self.alpha ** 2) / (1 - self.alpha)
+
+    def reset(self):
+        self.smoothed_action = None
+        return self.env.reset()
+
+    def step(self, action):
+        if self.smoothed_action is None:
+            self.smoothed_action = np.zeros_like(action)
+        self.smoothed_action = self.smoothing_coef * self.smoothed_action + (1 - self.smoothing_coef) * action
+        return self.env.step(self.smoothed_action)
 
 
 class DelayedRewardWrapper(gym.Wrapper):
@@ -191,6 +355,73 @@ class HistoryWrapper(gym.Wrapper):
         self.action_history = np.roll(self.action_history, shift=-action.shape[-1], axis=-1)
         self.action_history[..., -action.shape[-1] :] = action
         return self._create_obs_from_history(), reward, done, info
+
+
+class HistoryWrapperObsDict(gym.Wrapper):
+    """
+    History Wrapper for dict observation.
+
+    :param env: (gym.Env)
+    :param horizon: (int) Number of steps to keep in the history.
+    """
+
+    def __init__(self, env, horizon=5):
+        assert isinstance(env.observation_space.spaces["observation"], gym.spaces.Box)
+
+        wrapped_obs_space = env.observation_space.spaces["observation"]
+        wrapped_action_space = env.action_space
+
+        # TODO: double check, it seems wrong when we have different low and highs
+        low_obs = np.repeat(wrapped_obs_space.low, horizon, axis=-1)
+        high_obs = np.repeat(wrapped_obs_space.high, horizon, axis=-1)
+
+        low_action = np.repeat(wrapped_action_space.low, horizon, axis=-1)
+        high_action = np.repeat(wrapped_action_space.high, horizon, axis=-1)
+
+        low = np.concatenate((low_obs, low_action))
+        high = np.concatenate((high_obs, high_action))
+
+        # Overwrite the observation space
+        env.observation_space.spaces["observation"] = gym.spaces.Box(low=low, high=high, dtype=wrapped_obs_space.dtype)
+
+        super(HistoryWrapperObsDict, self).__init__(env)
+
+        self.horizon = horizon
+        self.low_action, self.high_action = low_action, high_action
+        self.low_obs, self.high_obs = low_obs, high_obs
+        self.low, self.high = low, high
+        self.obs_history = np.zeros(low_obs.shape, low_obs.dtype)
+        self.action_history = np.zeros(low_action.shape, low_action.dtype)
+
+    def _create_obs_from_history(self):
+        return np.concatenate((self.obs_history, self.action_history))
+
+    def reset(self):
+        # Flush the history
+        self.obs_history[...] = 0
+        self.action_history[...] = 0
+        obs_dict = self.env.reset()
+        obs = obs_dict["observation"]
+        self.obs_history[..., -obs.shape[-1] :] = obs
+
+        obs_dict["observation"] = self._create_obs_from_history()
+
+        return obs_dict
+
+    def step(self, action):
+        obs_dict, reward, done, info = self.env.step(action)
+        obs = obs_dict["observation"]
+        last_ax_size = obs.shape[-1]
+
+        self.obs_history = np.roll(self.obs_history, shift=-last_ax_size, axis=-1)
+        self.obs_history[..., -obs.shape[-1] :] = obs
+
+        self.action_history = np.roll(self.action_history, shift=-action.shape[-1], axis=-1)
+        self.action_history[..., -action.shape[-1] :] = action
+
+        obs_dict["observation"] = self._create_obs_from_history()
+
+        return obs_dict, reward, done, info
 
 
 class PlotActionWrapper(gym.Wrapper):